requirements.txt
asgiref==3.5.0
asn1crypto==1.5.1
autopep8==1.4.4
Django==4.0.6
django-allauth==0.39.1
django-cors-headers==3.13.0
django-countries==5.3.3
django-cloudinary-storage
django-crispy-forms==1.14.0
django-daraja==1.2.0
django-debug-toolbar==1.10.1
django-oso==0.26.0
gunicorn==20.1.0
oauthlib==3.0.1
Pillow==9.2.0
PyPDF3==1.0.6
python-bidi==0.4.2
python-decouple==3.1
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
requests-oauthlib==1.2.0
stripe
sqlparse==0.2.4
whitenoise==6.2.0
xhtml2pdf==0.2.8

